refactor: split typed api and extract project service core
