+++ src/deps/services/shared/typedCommandApi/state.js
@@ -1,0 +1,9 @@
+export const createStateCommandApi = (shared) => ({
+  getState: shared.getStateImpl,
+
+  getDomainState: shared.getDomainStateImpl,
+
+  async getEvents() {
+    return shared.getEventsImpl();
+  },
+});
